Moved dssp into separate module and separated original pydssp and my implementation
