+++ package/MDAnalysis/analysis/dssp/__init__.py
@@ -1,0 +1,8 @@
+__all__ = [
+    "DSSP",
+    "get_hbond_map",
+    "assign",
+    "translate",
+]
+
+from .dssp import DSSP, get_hbond_map, assign, translate
