+++ testsuite/MDAnalysisTests/analysis/test_dssp.py
@@ -2,19 +2,11 @@ import pytest
 import glob
 import MDAnalysis as mda
 
-from MDAnalysis.analysis.dssp import DSSP, translate, HAS_EINOPS
+from MDAnalysis.analysis.dssp import DSSP, translate
 from MDAnalysisTests.datafiles import DSSP as DSSP_FOLDER
 from MDAnalysisTests.datafiles import TPR, XTC
 
 
-@pytest.mark.skipif(HAS_EINOPS, reason="einops present")
-def test_without_einops():
-    u = mda.Universe(TPR, XTC).select_atoms('protein').universe
-    with pytest.raises(ImportError):
-        DSSP(u).run()
-
-
-@pytest.mark.skipif(not HAS_EINOPS, reason="einops present")
 @pytest.mark.parametrize("pdb_filename", glob.glob(f"{DSSP_FOLDER}/*.pdb.gz"))
 def test_file_guess_hydrogens(pdb_filename):
     u = mda.Universe(pdb_filename)
@@ -26,7 +18,6 @@ def test_file_guess_hydrogens(pdb_filename):
     assert answ == correct_answ
 
 
-@pytest.mark.skipif(not HAS_EINOPS, reason="einops present")
 def test_trajectory():
     u = mda.Universe(TPR, XTC).select_atoms('protein').universe
     run = DSSP(u).run(stop=10)
@@ -37,7 +28,6 @@ def test_trajectory():
     assert first_frame and last_frame and avg_frame
 
 
-@pytest.mark.skipif(not HAS_EINOPS, reason="einops present")
 def test_trajectory_with_hydrogens():
     u = mda.Universe(TPR, XTC).select_atoms('protein').universe
     run = DSSP(u, guess_hydrogens=False).run(stop=10)
